requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
coloredlogs==15.0.1
croniter==1.4.1
dateutils==0.6.12
deepdiff==6.3.1
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
lightning==2.0.5
lightning-cloud==0.5.37
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
onnx==1.14.0
ordered-set==4.1.0
packaging==23.1
protobuf==4.23.4
psutil==5.9.5
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.5
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
requests==2.31.0
rich==13.4.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
torch==2.0.1
torchmetrics==1.0.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
